Handle file skipping internally
The current solution for skipping files is to have a system utility such
as `ack` handle the searching for you.  This is not ideal because it
makes the documentation not as clear and ads a dependancies in some
cases.

Here we have the library handle skipping files, but simply not
transforming files that contain the tag