feat: Add support for a dry run option for ci environments #188 (#288)
Closes #188
Closes #169

This introduces a dry run option which rather than writing changes to the files instead returns an exit code of 1. This exit code can then be used in ci environments to fail a release/build due to out of date documentation.

Note this corrects the check to see if a TOC needs to be transformed as indicated in #169 and as such that pr should be closed.

Co-authored-by: Andrew Smith <andrewsmith@alumni.stanford.edu>